chore - reorganização de pastas por funcionalidade
