+++ src/main/java/br/edu/utfpr/pb/ext/server/auth/AuthController.java
@@ -1,4 +1,4 @@
-package br.edu.utfpr.pb.ext.server.controller;
+package br.edu.utfpr.pb.ext.server.auth;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
